app/requirements.txt
algoliasearch==2.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
praw==6.4.0
prawcore==1.0.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
update-checker==0.16
urllib3==1.25.6
websocket-client==0.56.0
Werkzeug==0.16.0
wikipedia==1.4.0
